pyproject.toml
[tool.poetry]
name = "kafka"
version = "0.1.0"
description = ""
authors = ["Kirill Nerush <nerush.kirill@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "^2.0.2"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
faker = "^20.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
